Move client content store to proxy package
